Update ILRepack
The ancient version of ILRepack we were using was broken due to a bug in Mono.Cecil that caused it to crash when running in a modern .NET context.
Even ignoring that, it chokes on modern resources embeddings.

The package we were using hasn't been updated since early 2019.

There is a fork that branched off in 2013 https://www.nuget.org/packages/ILRepack.Lib.MSBuild.Task but it doesn't work quite the same way (it tries to work automagically) and isn't a drop-in replacement.

As such this change just invokes the executable from the official distribution of ILRepack.

This change also makes it so that repacking is no longer done automatically during release builds. It'll be done explicitly by release automation.